feature authorization of superuser is done
