+++ dependencies.py
@@ -60,3 +60,24 @@ async def get_current_user_from_token(token:str = Depends(oauth2_scheme)):
     if user is None:
         raise credentials_exception
     return user
+
+
+async def get_current_admin_from_token(token:str = Depends(oauth2_scheme)):
+    credentials_exception = HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+    detail="Could not validate credentials")
+    adm_credentials_exception = HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+    detail="Could not validate admin credentials")
+    try:
+        payload = jwt.decode(token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM])
+        username: str = payload.get("sub")
+        if username is None:
+            raise credentials_exception
+        token_data = TokenData(username=username)
+    except JWTError:
+        raise credentials_exception
+    user = await getUserByUsername(username=token_data.username)
+    if user is None:
+        raise credentials_exception
+    if user.is_superUser==False:
+        raise adm_credentials_exception
+    return user
